+++ src/components/SearchFilters/index.tsx
@@ -1,0 +1,1 @@
+export { SearchFilters } from './SearchFilters';
