+++ src/test/__tests__/components/SearchControls.test.tsx
@@ -1,46 +1,44 @@
-import { fireEvent, render, screen } from '@testing-library/react';
+import { render } from '@testing-library/react';
 import { RecoilRoot } from 'recoil';
 import { getMockedImportedConstant } from '@src/test/__mocks__/common/constants/constants.mock';
 import * as FeatureConstants from '@common/constants/feature.constants';
 import { SearchControls } from '@components/SearchControls';
 
 const setSearchParams = jest.fn();
+const mockSearchFiltersComponent = <div data-testid="search-filters" />;
 
 jest.mock('react-router-dom', () => ({
   ...jest.requireActual('react-router-dom'),
   useSearchParams: () => [{}, setSearchParams],
 }));
+jest.mock('@components/SearchFilters', () => ({
+  SearchFilters: () => mockSearchFiltersComponent,
+}));
 
 describe('SearchControls', () => {
   const mockedSearchFiltersEnabled = getMockedImportedConstant(FeatureConstants, 'SEARCH_FILTERS_ENABLED');
-  mockedSearchFiltersEnabled(true);
 
-  beforeEach(() =>
-    render(
+  test('renders SearchFilters component', () => {
+    mockedSearchFiltersEnabled(true);
+
+    const { getByTestId } = render(
       <RecoilRoot>
         <SearchControls submitSearch={jest.fn} clearValues={jest.fn} />
       </RecoilRoot>,
-    ),
-  );
+    );
 
-  test('changes limiters', () => {
-    const initRadio = screen.getByRole('radio', { name: 'marva.allTime' });
-
-    expect(initRadio).toBeChecked();
-
-    fireEvent.click(screen.getByRole('radio', { name: 'marva.past12Months' }));
-
-    expect(initRadio).not.toBeChecked();
+    expect(getByTestId('search-filters')).toBeInTheDocument();
   });
 
-  test('adds and removes to the selection of limiters with multiselection option', () => {
-    const initCheckbox = screen.getByRole('checkbox', { name: 'marva.volume' });
-    expect(initCheckbox).not.toBeChecked();
+  test('does not render SearchFilters component', () => {
+    mockedSearchFiltersEnabled(false);
 
-    fireEvent.click(initCheckbox);
-    expect(initCheckbox).toBeChecked();
+    const { queryByTestId } = render(
+      <RecoilRoot>
+        <SearchControls submitSearch={jest.fn} clearValues={jest.fn} />
+      </RecoilRoot>,
+    );
 
-    fireEvent.click(initCheckbox);
-    expect(initCheckbox).not.toBeChecked();
+    expect(queryByTestId('search-filters')).not.toBeInTheDocument();
   });
 });
